Crate mithril_persistence

Shared datatypes and traits used to for the persistence layer of Mithril nodes

Modules

database

database module. This module contains queries and entities shared between all application types.

sqlite

SQLite module. This module provides a minimal yet useful Entity framework on top of SQLite with ability to perform any SQL query possible and hydrate results in Rust structs.

store

Define traits of StakeStorer.