mithril_persistence::database

Module record

Shared database records

Structs

BlockRangeRootRecord

Block range root record is the representation of block range with its merkle root precomputed.

CardanoBlockRecord

Cardano block record is the representation of a cardano block stored in the sqlite database.

CardanoBlockTransactionsRecord

Cardano block record is the representation of a cardano block stored in the sqlite database.

CardanoTransactionRecord

Cardano Transaction record is the representation of a cardano transaction.

StorableCardanoTransactionRecord

The record representation of a cardano transaction that can be stored in the database.