Module mithril_persistence::store

Define a generic way to store data with the Store Adapters, and the StakeStorer to store stakes.

Modules

adapter

Define a generic way to store data with the Store Adapter, with an adapter in memory and another sqlite.

Structs

StakeStore

A StakeStorer that use a StoreAdapter to store data.

Traits

StakeStorer

Represent a way to store the stake of mithril party members.

StorePruner

Implementing this trait will make store able to limit the number of the stored records by pruning them if a limit is set.