Module mithril_persistence::store::adapter

Define a generic way to store data with the Store Adapter, with an adapter in memory and another sqlite.

Structs

DumbStoreAdapter

A StoreAdapter that store one fixed data record, for testing purpose.

FailStoreAdapter

A StoreAdapter which always fails, for testing purpose.

MemoryAdapter

A StoreAdapter that store data in memory.

SQLiteAdapter

Store adapter for SQLite3

SQLiteResultIterator

Iterator over SQLite adapter results.

Enums

AdapterError

StoreAdapter related errors

Traits

StoreAdapter

Represent a way to store Key/Value pair data.