mithril_persistence

Module database

database module. This module contains queries and entities shared between all application types.

Modules

cardano_transaction_migration

Migration module for cardano transactions store

record

Shared database records

repository

Shared database repositories

Structs

DatabaseVersion

Entity related to the db_version database table.

DatabaseVersionChecker

Struct to perform application version check in the database.

GetDatabaseVersionQuery

Query to get DatabaseVersion entities.

Hydrator

Helper struct to hydrate common data.

SqlMigration

Represent a file containing SQL structure or data alterations.

UpdateDatabaseVersionQuery

Query to UPSERT DatabaseVersion entities.

Enums

ApplicationNodeType

Application using a database

Type Aliases

DbVersion

Database version.