mithril_persistence
Module sqlite

SQLite module. This module provides a minimal yet useful Entity framework on top of SQLite with ability to perform any SQL query possible and hydrate results in Rust structs.

Modules

error

Helpers to handle SQLite errors

Structs

ConnectionBuilder

Builder of SQLite connection

EntityCursor

Database query result Iterator wrapper. This wrapper allows to call entity hydration for each extracted result.

Projection

Projection is a definition of field mapping during a query. Fields come from one or several source structures (can be tables, views or sub queries) and are mapped to a Query as output.

ProjectionField

Each projection field is defined by

SourceAlias

Handful tool to store SQL source aliases.

SqliteCleaner

The SqliteCleaner is responsible for cleaning up databases by performing tasks defined in SqliteCleaningTask.

SqliteConnectionPool

Pool of Sqlite connections

SqlitePooledConnection

SqliteConnection wrapper for a pooled connection

Transaction

Sqlite transaction wrapper.

WhereCondition

Where condition builder.

Enums

ConnectionOptions

Options to apply to the connection

HydrationError

SqLite hydration error

SqliteCleaningTask

Tasks that can be performed by the SqliteCleaner

Traits

ConnectionExtensions

Extension trait for the SqliteConnection type.

GetAllCondition

Get all condition builder.

Query

Define a query to perform on database and return iterator of a defined entity.

SqLiteEntity

How to hydrate an entity from a SQLite result row

Type Aliases

SqliteConnection

Type of the connection used in Mithril